Removing call to parent class constructor from WPcom_JS_Concat::__construct
Since the `WP_Scripts::__construct` has either already been called outside the WPcom_JS_Concat class or is called from within the `WPcom_JS_Concat::__construct` during the initialisation of the `old_scripts` variable, we should not call the parent class consturctor again, as it fires the `wp_default_scripts` action again.

The `wp_default_scripts` action is designed for 2 runs, before and after the `init` action. If it's run after the `init` action, localisations are being populated and calling that action for the second time after the `init` action leads to doubling the localisation strings.