Simplify realpath to fix subdirectory installs
If this is a subdirectory site, we need to strip off the subdir from the URL before trying to determine its realpath.

This is because in a multisite install, the subdir is virtual and therefore not needed in the path(For `example.com/subdir/file.css`, `file.css` is actually served from `/`).

In a single-site subdir install, the subdir is included in the ABSPATH and therefore ends up duplicated (ABSPATH: /var/www/wp | URL: example.com/wp/file.css).