Enhancing the regular expression used for replacing relative paths to absolute ones the way it really only excluded the ones starting with a hash.
The regular expression introduced in 2e29e538c7a0fb9ad73f68814e992fc5d611eef1 is excluding not only URLs starting with a hash, but also any URLs containg a hash, which is not correct and prevents those from being properly served.

The negative lookbehind used for catching a hash has been extended the way it looks for a hash preceding either with `(`, `"` or `'` in order to match only URLs starting with a hash, not URLs containing a hash.

This fixes test cases introduced in previous commits.