Implement 'js_concat_script_attributes' to allow users to add extra tag attributes, the filter gets passed an empty array and has to return an associative or regular array with attributes as keys, additional arguments for the filter are: $href - current batch url, $js_array - additional data for current batch of scripts.
This filter runs on each batch of files; extra arguments are needed to better control attributes being assigned to a resulting tag.

For example, we want to set async/defer for either a specific group of files being concatenated or maybe we want to do a blanket change for every resulting script tag.

Fixes #33, supersedes #36