Disable concat for inline translations
WordPress 5.0 added a new way to deliver translations for the editor.  Translations are stored in JSON files and loaded inline through `print_translations`.  Our JS concatenation currently doesn’t account for this and so the translations fail to load.

This follows the same approach that @dchymko did to prevent inline scripts from concatenating.