Subdirectory support
If we're in a subdirectory context, use that as the root. We can't assume that the root serves the same content as the subdir.

Previously, a request to `/subdir/_static/??...` would replace relative
paths (like `../path/to/file.svg`) to `/path/to/file.svg`. Now, we set
it to `/subdir/path/to/file.svg`.