Properly handle groups assigned for enqueing in footer
This is a follow-up of the 4bc494e2f4a6b18aed123ee1e639ddba7f23697b which inproved printing of extra JS data for groups (scripts with empty src).

Such a group can be set for printing/loading in footer, but since we are handling such cases later in the code, and since we `continue` in the foreach cycle inside the conditional checking for groups, this never happens.

This changeset fixes the issue by performing the in_footer check before empty groups.